pyproject.toml
[project]
name = "neural-operators"
version = "0.1.0"
description = "Neural operator library: FNO and DeepONet for PDE operator learning"
requires-python = ">=3.12,<3.14"
dependencies = [
    "torch==2.5.1",
    "nvidia-cublas-cu12",
    "nvidia-cuda-cupti-cu12",
    "nvidia-cuda-nvrtc-cu12",
    "nvidia-cuda-runtime-cu12",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12",
    "nvidia-curand-cu12",
    "nvidia-cusolver-cu12",
    "nvidia-cusparse-cu12",
    "nvidia-nccl-cu12",
    "nvidia-nvjitlink-cu12",
    "nvidia-nvtx-cu12",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "scipy-stubs~=1.15.3",
    "path>=17.1.1",
    "neuraloperator>=2.0.0",
    "pyarrow>=24.0.0",
    "pandas>=2.3.3",
    "pyvista[all]>=0.47.3",
]

[dependency-groups]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.13.0",
]

[tool.uv.sources]
torch = { index = "pytorch-cu121" }

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neural_operators"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501", "E701", "E702", "UP046"]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["E402"]  # sys.path.insert before imports required in standalone scripts

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: fast tests using in-memory stubs, no real data files",
    "integration: tests that require real data files or trained checkpoints",
]
